Initial mockup of layout/view editing
Implement an initial mockup of layout editing. This includes jQuery UI support to drag and resize form fields within a parent layout container. The fields' positions are initialized using Javascript. Changes to fields' position and size are also tracked. The next step is to load and save layout information to the datastore.