Setup an initial, prototype page for editing records
This includes refactoring the Javascript written for editing a layout to have a reusable function to initialize the layout from a datastore. The only difference is, when editing records, the fields are editable and changes to values are saved back to the datastore. Some of the other common/reusable type code previously written for editing a layout has also been moved to separate javascript files.