Consolidate all the functionality for calculated field evaluation/calculation into calcFieldEqnEval.go
For purposes of maintainability and modularity, this organization seems to work better than spreading the code into the other source files for Field and Record. This keeps field.go and record.go focused solely on saving and retrieving the information from the datastore, but moves code related to equation evaluation together in calcFieldEqnEval.go.