Initial API support to create calculated fields and support in the web client to view calculated fields
This included separating the constructors for Field structs into one for calculated fields and one for regular fields, with an internal function to do the "raw" construction. The validation function used to validate a value being set vs the field it is associated with has also been enhanced to support contexts where calculated fields are allowed or not.