Initial refactoring to group CSS, HTML template and Javascript files together for source code changes
There will be numerous more visual elements, each with an HTML template and closely related Javascript file to manage the DOM elements defined by the template. In the interests of "what changes together, stays together", it makes sense to group these files together in the source tree. It may become necessary to copy these files out to different folders to build the final project (e.g. when minimizing Javascript). Since the code will be managed 95% of the time for source code maintenance, it make sense to put files together when their purpose and functionality is very closely coupled and generally have a similar dependencies.