Refactor the project file and directory organization for the web UI - closely related Javascript, HTML, Golang, and CSS files are now grouped together.
This gives precedence to put files with a tight coupling together in the same folder. These files all have similar dependencies to other directories/packages, so the folder layout takes into account dependencies as well. The goal is to make each package as autonomous and stand-alone as possible, so that maintenance can occur in a tightly grouped set of files & folders, rather than across many files and folders. Some additional scripting might be necessary for deployment, but the source code is being organized for humans not machines.