Refactor the project file and folder organization to eliminate build conflicts when building the GAE project
See the README file in googleAppEng/ for details.

This commit also includes a WIP refactoring of the TextBox server code and client code to support modular editing of TextBoxes and other objects. The initial client code was hard-wired for text boxes, so refactoring is needed to make the Javascript more reusable with check boxes and other form elements.