Update the textBox package to place emphasis back on MVC and plug-in type architecture for form objects
Since the textBox is it's own package, the files can be named to represent where they fit in the architecture. In particular:

* dataModel.go contains methods to create new, update existing, get all (and soon delete) TextBoxes. Functions in this file wrap datastoreWrapper functions for a lower-level access to the datastore.
* httpAPI.go - includes both the http route definitions and functions called for each route. Notably, the routes are defined autonomously in the init() function, so the package is standalone vs other similar packages.
* updateProps.go - this is where functions to reposition, resize and perform other property updates on the text box reside. This package has a similar "property updater" design as other packages.

With this update the textBox package is probably a reference design for different packages on the server. The code is relatively tight and robust and should be easy to maintain.