WIP on form checkbox webui code - The dialog to create new checkboxes now validates the information and passes it back to a saveNewCheckBox() function.
The wizard dialog handling was also enhanced to support a "form info" object which is returned from each dialog panel's init function. This form info is used after the dialog exits to retrieve field values. This is necessary because the DOM element ID's themselves can be based upon a prefix to create multiple instances of the same dialog panel.