Enhance tokenizer to optionally tokenize both whitespace and comments
This is needed so the tokenizer can be used to preprocess calculated field multi-line formulas to replace field references with unique field IDs before storing the source into the datastore (and vice-versa coming out of the datastore). When the formula is being parsed for real calculations, whitespace and comments can still be ignored.