Implement a pre-processor for substituting field references with field IDs (and vice-versa)
See the comment in preprocessor.go for details - basically this is needed so the user can put "field reference names" in formulas/equations, but store the formulas using permanent field IDs. The preprocessing also preserves comments and whitespace, so if the user edits the formula again, comments and whitespace are preserved. The next step is to transition the actual equation storage and processing to support pre-processing.