Refactor handling of parent and child entity IDs and clean up the datastore wrapper package
In preparation for further implementation on the data model, the existing code for passing around entity IDs and interfacing with the data store has been refactored. In particular, instead of passing around both a parent and child ID, a single encoded child ID is now passed back from the datastoreWrapper package. This encoded child ID is a concatenation of an encoded version of both the parent and child datastore keys. This simplifies the code in both the server and web ui, since only a single opaque ID is passed around. The interface to the server is also simplified since clients aren't exposed to "object IDs" and "parent IDs"; they just receive a single opaque string ID to work with.

The only downside to these changes is some of the more generic code for updating properties now needs to be specialized per entity type. In particular, when there was a generic header with a struct including a parent ID and object ID, this header could be used by all the property updaters for different entities. However, this also gave the API into the server a more "programmatic" implementation which exposed some of the server-side details of unique ID handling. Although there's now a little more code per entity, the interface is more compact with more like "tableID: ENCODED_TABLE_ID" instead of "uniqueID { parentID: DATABASE_ID, objectID: TABLE_ID }" (where the upper case identifiers refer to encoded/opaque IDs).

As part of this refactoring, the datastoreWrapper package has also been updated to make some of the functions which should have been private to the package private. This resulted in updates to several clients of the package to use the preferred wrapper functions rather than the private functions. This package needs a little bit more work to further hide some of its implementation (e.g. encoding/decoding of keys), but overall the API to this package is pretty stable and robust at this point.