Refactor the datastoreWrapper package to start using the native Encode() and DecodeKey() function instead of encoded ints
There was a problem with storing and retrieving descendent entities when the opaque IDs for keys were returned as a base36 encoding of the integer key. To store and retrieve a "grandchild" entity requires fully-defined keys to be instantiated up the entire ancestry tree. On the other hand the native Encode() and DecodeKey() functions serializes a "fully defined" key, so that an entity can be updated or retrieved later without the entire ancestry tree being known. This results in much simpler and compact code.