Include table ID as a Javascript global in pages to view and design the form
To make records and fields a child of a specific table, they also need to be retrieved based upon a specific table. A record and field will have the same parent table ID as it's parent, so it suffices when loading a form to include the form's parent table ID as the table ID with which to retrieve the matching records and fields.