Start migrating to a time series for record updates.
This update includes definition of a cell_updates table. The values are currently being written to this table, alongside the old table for the values stored in the record itself. More work is needed to map the values in the cell_updates table back to values which can be shown to the user.