Move lower-level functionality to write record updates back into the record package
The API for record updates is still in the higher level recordUpdate package. The recordUpdate package or something similar needs to be a controller package to coordinate writes to the datastore, but also trigger calculated field updates, re-mapping of record values, etc. This needs further design w.r.t. using some kind of listener pattern in the record package to be notified of record updates, or route record updates through a controller package.