Work in progress implementation of drag and drop for dashboard & form components, including support to create new columns or rows.
After struggling to get jQuery UI sortable working, a custom implementation is now in place which still uses jQuery UI draggable, but has custom logic for inserting placeholders while components are being dragged around in design mode. jQuery UI sortable was not a good fit for this type of nested sorting.

The drop behavior still needs to be reintegrated, as well as new behavior to create new rows or columns when the component is dropped on a placeholder for a new row or column.