In note editor and caption form components, set the target="_blank" attribute on all links.
This ensures new links will open in a new window or tab, rather than exiting the application.

The editing behavior for notes was changed to require clicking on an edit button, which enables the editor. When the blur event is received from the editor, the changed value is written. Manually enabling editing ensures users can still visit links within the note when not editing the note.