Eliminate reading & writing of calculated values to the database. Just re-calculating the values is faster and has less code complexity.
For a sample data set, before the optimization to read all the values at once and to eliminate writing calculated values, it took 486 msec to re-compute the data set. Afterwards, it takes about 80 sec (a savings of almost 85%). More work is needed to use an LRU cache so that values aren't recomputed when they're only being viewed (e.g., Dashboards).