Initial algorithm to generate alerts when record updates are submitted.
The basic algorithm to generate or re-generate a list of alert notifications has been implemented. It's only working for a single type of alert and one record at a time. More design & implementation is needed to save the alert notifications and handle cases where all the alerts need to be updated at once, such as when the alerts themselves change.