Fix vertical resizing of the table view
Previously a plug-in was used to vertically resize the table view. However, the results were sporadic and scroll bars would appear and disappear depending upon the resizing behavior. This approach simply adjust the page size (number of rows per page) based upon the height of the surrounding container, header and footer.