Work in progress on a "main window" layout
The intent is to keep this layout fixed on the page at all times, then dynamically load the content within viewports (lists, dashboards, forms). The menus and sidebars will then not have to be loaded each time a new list or dashboard is selected. This also allows for "lazy loading" of javascript libraries which only pertain to lists or dashboards.